HtmlBridge Phase 3 ratchet: de-list DomBridge.Serialization.cs (split SVG-zoom partial)
Peel the cohesive SVG zoom-serialization attribute-scaling cluster out of
DomBridge.Serialization.cs into a sibling partial DomBridge.Serialization.SvgZoom.cs
(300 lines): ApplyZoomSerializationSvgAttributes and the presentation/geometry
scalers, the SVG font-relative unit resolution helpers, the four SVG unit sets,
and the three [GeneratedRegex] point/path/font-shorthand patterns. Only the
ApplyZoomSerializationStyles entry call stays behind; the now-unused
System.Text.RegularExpressions using is dropped from it.

Pure partial-class relocation — no signature, accessibility, or logic change,
so behaviour-identical by construction. Serialization.cs drops 951 -> 668 lines,
under the 750-line guard, so its OversizedFileExemptions entry is removed and the
stale-exemption check now enforces it stays under the limit. Debt list:
seven -> six files.

Broiler.HtmlBridge.Dom builds clean; HtmlBridgeArchitectureGuardTests 14/14 green
(0 offenders / 0 stale exemptions); SVG-serialization / zoom coverage
(SvgDomAndCrossDocTests, SvgDomDynamicContentTests, SharedGeometryZoomSizeTests,
Acid3SvgAndParsingRegressionTests, 81 tests) green.
